A few compile errors that weren't showing up in vscode.  Tried setting the c standard to old c89 but it didn't seem to matter.  Surprisingly difficult to find online what the C standard actually was back in the 2003 visual studio.

+++ UtilityLib/GraphicsDevice.h
@@ -20,7 +20,7 @@ extern D3DTexture	*GD_CreateTexture(GraphicsDevice *pGD,
 
 //shader creation stuff
 extern DWORD	GD_CreateVertexShader(GraphicsDevice *pGD,
-					const DWORD *pDecl,	const BYTE *pCodeBytes);
+					const DWORD *pDecl,	const DWORD *pCodeBytes);
 extern DWORD	GD_CreatePixelShader(GraphicsDevice *pGD,
 					const D3DPIXELSHADERDEF *pCodeBytes);
